fix build errors due to errors reexpo in the context of worker
