+++ rust/shell/src/api/mod.rs
@@ -1,5 +1,6 @@
 mod place;
 mod post;
+mod error;
 
 use crate::db::DbPool;
 use crate::error::Error;
