+++ rust/shell/src/error.rs
@@ -1,103 +1,130 @@
 #[derive(Debug)]
 pub enum Error {
-    DieselConnectionError(diesel::result::ConnectionError),
-    DieselQueryError(diesel::result::Error),
     DotEnvyError(dotenvy::Error),
-    FromUTF8Error(std::string::FromUtf8Error),
-    GeoJsonError(geojson::Error),
-    HMacError(hmac::digest::InvalidLength),
-    JwtError(jwt::Error),
-    LapinError(lapin::Error),
-    R2D2Error(r2d2::Error),
-    RedisError(redis::RedisError),
-    SendError(std::sync::mpsc::SendError<(String, String)>),
-    SerializeError(serde_json::Error),
     StdEnvVarError(std::env::VarError),
 
-    AuthError(AuthError),
-    HttpError(HttpError),
-    GeoJSONSerdeError(GeoJSONSerdeError),
+    #[cfg(feature="db")]
+    DieselConnectionError(diesel::result::ConnectionError),
+    #[cfg(feature="db")]
+    DieselQueryError(diesel::result::Error),
 
-    NotFound,
-    Forbidden,
-    Unauthorized,
-    Conflict,
-    Gone,
-    NotModified,
-    PreconditionFailed,
+    #[cfg(feature="api")]
+    HMacError(hmac::digest::InvalidLength),
+    #[cfg(feature="api")]
+    JwtError(jwt::Error),
+    #[cfg(feature="api")]
+    AuthError(AuthError),
+    #[cfg(feature="api")]
+    HttpError(HttpError),
+
+    #[cfg(any(feature="rmqpub", feature="rmqsub", feature="redis", feature="db"))]
+    R2D2Error(r2d2::Error),
+
+    #[cfg(any(feature="rmqpub", feature="rmqsub"))]
+    LapinError(lapin::Error),
+    #[cfg(feature="rmqsub")]
+    GeoJsonError(geojson::Error),
+    #[cfg(feature="rmqsub")]
+    GeoJSONSerdeError(GeoJSONSerdeError),
+    #[cfg(feature="rmqsub")]
+    FromUTF8Error(std::string::FromUtf8Error),
+    #[cfg(feature="rmqpub")]
+    SendError(std::sync::mpsc::SendError<(String, String)>),
+    
+    #[cfg(feature="redis")]
+    RedisError(redis::RedisError),
+
+    #[cfg(any(feature="rmqpub", feature="api"))]
+    SerializeError(serde_json::Error),
 
     Error, // Generic error
 }
+#[cfg(feature="api")]
 impl From<AuthError> for Error {
     fn from(value: AuthError) -> Self {
         Error::AuthError(value)
     }
 }
+#[cfg(feature="api")]
 impl From<HttpError> for Error {
     fn from(value: HttpError) -> Self {
         Error::HttpError(value)
     }
 }
+#[cfg(feature="rmqsub")]
 impl From<GeoJSONSerdeError> for Error {
     fn from(value: GeoJSONSerdeError) -> Self {
         Error::GeoJSONSerdeError(value)
     }
 }
+
+#[cfg(feature="db")]
 impl From<diesel::result::ConnectionError> for Error {
     fn from(value: diesel::result::ConnectionError) -> Self {
         Error::DieselConnectionError(value)
     }
 }
+#[cfg(feature="db")]
 impl From<diesel::result::Error> for Error {
     fn from(value: diesel::result::Error) -> Self {
         Error::DieselQueryError(value)
     }
 }
+
 impl From<dotenvy::Error> for Error {
     fn from(value: dotenvy::Error) -> Self {
         Error::DotEnvyError(value)
     }
 }
+#[cfg(feature="rmqsub")]
 impl From<std::string::FromUtf8Error> for Error {
     fn from(value: std::string::FromUtf8Error) -> Self {
         Error::FromUTF8Error(value)
     }
 }
+#[cfg(feature="rmqsub")]
 impl From<geojson::Error> for Error {
     fn from(value: geojson::Error) -> Self {
         Error::GeoJsonError(value)
     }
 }
+#[cfg(feature="api")]
 impl From<hmac::digest::InvalidLength> for Error {
     fn from(value: hmac::digest::InvalidLength) -> Self {
         Error::HMacError(value)
     }
 }
+#[cfg(feature="api")]
 impl From<jwt::Error> for Error {
     fn from(value: jwt::Error) -> Self {
         Error::JwtError(value)
     }
 }
+#[cfg(any(feature="rmqpub", feature="rmqsub"))]
 impl From<lapin::Error> for Error {
     fn from(value: lapin::Error) -> Self {
         Error::LapinError(value)
     }
 }
+#[cfg(any(feature="rmqpub", feature="rmqsub", feature="redis", feature="db"))]
 impl From<r2d2::Error> for Error {
     fn from(value: r2d2::Error) -> Self {
         Error::R2D2Error(value)
     }
 }
+#[cfg(feature="redis")]
 impl From<redis::RedisError> for Error {
     fn from(value: redis::RedisError) -> Self {
         Error::RedisError(value)
     }
 }
+#[cfg(feature="rmqpub")]
 impl From<std::sync::mpsc::SendError<(String, String)>> for Error {
     fn from(value: std::sync::mpsc::SendError<(String, String)>) -> Self {
         Error::SendError(value)
     }
 }
+#[cfg(any(feature="rmqpub", feature="api"))]
 impl From<serde_json::Error> for Error {
     fn from(value: serde_json::Error) -> Self {
         Error::SerializeError(value)
@@ -111,11 +138,52 @@ impl From<std::env::VarError> for Error {
 
 impl std::fmt::Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("shell error")
+        match &self {
+            Error::DotEnvyError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            Error::StdEnvVarError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+
+            #[cfg(feature="db")]
+            Error::DieselConnectionError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="db")]
+            Error::DieselQueryError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+
+            #[cfg(feature="api")]
+            Error::HMacError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="api")]
+            Error::JwtError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="api")]
+            Error::AuthError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="api")]
+            Error::HttpError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+
+            #[cfg(any(feature="rmqpub", feature="rmqsub", feature="redis", feature="db"))]
+            Error::R2D2Error(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            
+            #[cfg(any(feature="rmqpub", feature="rmqsub"))]
+            Error::LapinError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="rmqsub")]
+            Error::GeoJsonError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="rmqsub")]
+            Error::GeoJSONSerdeError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="rmqsub")]
+            Error::FromUTF8Error(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+            #[cfg(feature="rmqpub")]
+            Error::SendError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+
+            #[cfg(feature="redis")]
+            Error::RedisError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+
+            #[cfg(any(feature="rmqpub", feature="api"))]
+            Error::SerializeError(error) => f.write_str(format!("shell error {:?}", error).as_str()),
+
+
+            Error::Error => f.write_str("shell error"),
+        }
     }
 }
 impl std::error::Error for Error {}
 
+#[cfg(feature="api")]
 #[derive(Debug)]
 pub enum AuthError {
     NoAuthorizationHeader,
@@ -123,12 +191,21 @@ pub enum AuthError {
     NoSubjectClaim,
 }
 
+#[cfg(feature="api")]
 #[derive(Debug)]
 pub enum HttpError {
     NoIfMatchHeader,
     NoIfNoneMatchHeader,
+    NotFound,
+    Forbidden,
+    Unauthorized,
+    Conflict,
+    Gone,
+    NotModified,
+    PreconditionFailed,
 }
-#[derive(Debug)]
+
+#[derive(Debug, )]
 pub enum GeoJSONSerdeError {
     NoProperties,
     NoGeometry,
@@ -136,107 +213,14 @@ pub enum GeoJSONSerdeError {
     MissingProperty(String),
     InvalidPropertyType(String),
 }
-pub type ResponseError = rocket::response::status::Custom<String>;
-
-impl From<Error> for ResponseError {
-    fn from(value: Error) -> Self {
-        match value {
-            Error::AuthError(e) => rocket::response::status::Custom(
-                rocket::http::Status::Unauthorized,
-                format!("error: {:?}", e).to_string(),
-            ),
-            Error::HttpError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ),
-            Error::GeoJSONSerdeError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ),
-            Error::JwtError(e) => rocket::response::status::Custom(
-                rocket::http::Status::Unauthorized,
-                format!("error: {:?}", e).to_string(),
-            ),
-            Error::DotEnvyError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::StdEnvVarError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::DieselConnectionError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::DieselQueryError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::FromUTF8Error(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::GeoJsonError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::HMacError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::R2D2Error(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::RedisError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::LapinError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            // map every adapters error to 500
-            Error::SendError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::SerializeError(e) => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                format!("error: {:?}", e).to_string(),
-            ), // map every adapters error to 500
-            Error::NotFound => rocket::response::status::Custom(
-                rocket::http::Status::NotFound,
-                "not found".to_string(),
-            ),
-            Error::Forbidden => rocket::response::status::Custom(
-                rocket::http::Status::Forbidden,
-                "forbiddon".to_string(),
-            ),
-            Error::Conflict => rocket::response::status::Custom(
-                rocket::http::Status::Conflict,
-                "conflict".to_string(),
-            ),
-            Error::Gone => {
-                rocket::response::status::Custom(rocket::http::Status::Gone, "gone".to_string())
-            }
-            Error::Unauthorized => rocket::response::status::Custom(
-                rocket::http::Status::Unauthorized,
-                "unauthorized".to_string(),
-            ),
-            Error::NotModified => rocket::response::status::Custom(
-                rocket::http::Status::NotModified,
-                "not modified".to_string(),
-            ),
-            Error::PreconditionFailed => rocket::response::status::Custom(
-                rocket::http::Status::PreconditionFailed,
-                "precondition failed".to_string(),
-            ),
-            Error::Error => rocket::response::status::Custom(
-                rocket::http::Status::InternalServerError,
-                "error".to_string(),
-            ),
+impl std::fmt::Display for GeoJSONSerdeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match &self {
+            GeoJSONSerdeError::NoProperties => write!(f, "missing 'properties' field"),
+            GeoJSONSerdeError::NoGeometry => write!(f, "missing 'geometry' field"),
+            GeoJSONSerdeError::InvalidGeometryType => write!(f, "invalid geometry type"),
+            GeoJSONSerdeError::MissingProperty(p) => write!(f, "missing required property '{}'", p),
+            GeoJSONSerdeError::InvalidPropertyType(p) => write!(f, "invalid type for property '{}'", p),
         }
     }
 }
